+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **Medium Impact Changes**
 - **Other Features**
   - [spiral/queue] Added the ability to pass additional data in the `context` parameter in the job handlers.
+  - [spiral/queue] Added new options `priority`, `autoAck`, `headers` in the `Spiral\Queue\Options`.
 
 ## v3.1.0 - 2022-09-29
 - **Other Features**